Update Crusade Unit Form
Crusade View => MulipleStream View model
UpdateCrusadeUnitView => done

+++ lib/presentation/views/crusade_view/crusade_view_model.dart
@@ -8,11 +8,14 @@ import 'package:wh40k_crusader/app/locator.dart';
 import 'package:wh40k_crusader/data_models/crusade_data_model.dart';
 import 'package:wh40k_crusader/data_models/crusade_unit_data_model.dart';
 import 'package:wh40k_crusader/routing/routes.dart';
+import 'package:wh40k_crusader/routing/routing_args.dart';
 import 'package:wh40k_crusader/services/firestore_service.dart';
 
 enum CrusadeViewModelState { showRoster, showEditForm }
+const String kCrusadeStream = 'crusade-stream';
+const String kRosterStream = 'roster-stream';
 
-class CrusadeViewModel extends BaseViewModel {
+class CrusadeViewModel extends MultipleStreamViewModel {
   CrusadeDataModel crusade;
 
   final _db = locator<FirestoreService>();
@@ -20,10 +23,19 @@ class CrusadeViewModel extends BaseViewModel {
   final _dialogService = locator<DialogService>();
   final editCrusadeValuesFormKey = GlobalKey<FormBuilderState>();
 
-  List<CrusadeUnitDataModel> roster = [];
   CrusadeViewModelState _state = CrusadeViewModelState.showRoster;
   CrusadeViewModelState get state => _state;
   CrusadeViewModel(this.crusade);
+
+  @override
+  Map<String, StreamData> get streamsMap => {
+        kCrusadeStream: StreamData<CrusadeDataModel>(
+            _db.listenToCrusadeRealTime(crusadeUID: crusade.documentUID!),
+            onData: (data) => crusade = data),
+        kRosterStream: StreamData<List<CrusadeUnitDataModel>>(
+            _db.listenToCrusadeRoster(crusadeUID: crusade.documentUID!),
+            transformData: _orderRoster),
+      };
 
   setState(CrusadeViewModelState state) {
     setBusy(true);
@@ -42,31 +54,43 @@ class CrusadeViewModel extends BaseViewModel {
     notifyListeners();
   }
 
-  getCrusadeInfo() async {
-    // TODO consider why the form doesn't update on unit added to roster.
-    logger.i("Getting new data for Crusade: ${crusade.documentUID}");
-    CrusadeDataModel? newData = await _db.getCrusadeByUID(crusade.documentUID!);
-    roster = await _db.getRoster(crusadeUID: crusade.documentUID!);
-
-    if (newData != null) {
-      logger.i("New data for Crusade: ${crusade.documentUID}");
-      logger.wtf("${newData.toJSONString()}");
-      crusade = newData;
-      // _updateCrusadeFormFieldsValues();
-    }
-    _orderRoster();
-    notifyListeners();
-  }
+  // getCrusadeInfo() async {
+  //   // TODO consider why the form doesn't update on unit added to roster.
+  //   logger.i("Getting new data for Crusade: ${crusade.documentUID}");
+  //   CrusadeDataModel? newData = await _db.getCrusadeByUID(crusade.documentUID!);
+  //   roster = await _db.getRoster(crusadeUID: crusade.documentUID!);
+  //
+  //   if (newData != null) {
+  //     logger.i("New data for Crusade: ${crusade.documentUID}");
+  //     logger.wtf("${newData.toJSONString()}");
+  //     crusade = newData;
+  //     // _updateCrusadeFormFieldsValues();
+  //   }
+  //   _orderRoster();
+  //   notifyListeners();
+  // }
 
   dropUnitFromCrusadeRoster(CrusadeUnitDataModel unitToDrop) async {
-    await _db.deleteRemoveUnitFromRoster(
-        crusade: crusade, unitToDelete: unitToDrop);
-    await getCrusadeInfo();
+    DialogResponse? response = await _dialogService.showConfirmationDialog(
+      title: 'Delete Unit',
+      description: 'Are you sure you want to delete this unit?',
+      confirmationTitle: 'Cancel',
+      cancelTitle: 'Delete',
+    );
+
+    logger.i('Dialog Response ${response?.confirmed}');
+
+    if (!response!.confirmed) {
+      await _db.deleteRemoveUnitFromRoster(
+          crusade: crusade, unitToDelete: unitToDrop);
+    }
+
+    // await getCrusadeInfo();
   }
 
   _deleteRoster() async {
     await _db.deleteRoster(crusade);
-    await getCrusadeInfo();
+    // await getCrusadeInfo();
   }
 
   Future _deleteCrusadeDocument() async {
@@ -123,7 +147,7 @@ class CrusadeViewModel extends BaseViewModel {
     );
 
     await _db.updateCrusade(updatedData);
-    await getCrusadeInfo();
+    // await getCrusadeInfo();
   }
 
   navigateToCreateNewUnitForm() {
@@ -131,7 +155,14 @@ class CrusadeViewModel extends BaseViewModel {
         arguments: crusade);
   }
 
-  _orderRoster() {
+  navigateToUpdateCrusadeUnitView(CrusadeUnitDataModel unit) {
+    _navigationService.navigateTo(
+      rNavigationRoutes.UpdateUnitRoute,
+      arguments: UpdateCrusadeUnitRouteArgs(crusade, unit),
+    );
+  }
+
+  _orderRoster(List<CrusadeUnitDataModel> roster) {
     List<CrusadeUnitDataModel> hqs = [];
     List<CrusadeUnitDataModel> elites = [];
     List<CrusadeUnitDataModel> troops = [];
@@ -140,7 +171,6 @@ class CrusadeViewModel extends BaseViewModel {
     List<CrusadeUnitDataModel> heavySupport = [];
     List<CrusadeUnitDataModel> flyer = [];
     List<CrusadeUnitDataModel> low = [];
-    List<CrusadeUnitDataModel> sc = [];
     // Sort in Descending Experience Order
     roster.sort((b, a) => a.experience.compareTo(b.experience));
 
@@ -168,16 +198,12 @@ class CrusadeViewModel extends BaseViewModel {
       } else if (element.battleFieldRole ==
           CrusadeUnitDataModel.battleFieldRoles[7]) {
         low.add(element);
-      } else if (element.battleFieldRole ==
-          CrusadeUnitDataModel.battleFieldRoles[8]) {
-        sc.add(element);
       } else {
         logger.wtf("We have a unit with an unknown battlefield role");
       }
     });
 
     List<CrusadeUnitDataModel> orderedRoster = [];
-    sc.forEach((element) => orderedRoster.add(element));
     hqs.forEach((element) => orderedRoster.add(element));
     elites.forEach((element) => orderedRoster.add(element));
     troops.forEach((element) => orderedRoster.add(element));
@@ -187,6 +213,8 @@ class CrusadeViewModel extends BaseViewModel {
     flyer.forEach((element) => orderedRoster.add(element));
     low.forEach((element) => orderedRoster.add(element));
 
-    roster = orderedRoster;
+    roster.clear();
+    orderedRoster.forEach((element) => roster.add(element));
+    notifyListeners();
   }
 }
